차트 그래프 나옴(다만 창을 벗어나는 문제 있음), 약간  버그 수정

+++ mit/src/main/java/com/itbank/service/StoreService.java
@@ -51,6 +51,7 @@ public class StoreService {
       
       if (files.get(0).getSize() == 0) {
          dto.setPdimg("");
+         dto.setViewimg("");
          return dao.getInsertItem(dto);
       }
       
@@ -146,15 +147,17 @@ public class StoreService {
 
    public int delete(int idx) {
       StoreDTO dto = selectOne(idx);
-      String newdir = uploadPath + "\\" + dto.getPdcode() + dto.getPdwriter();
-      File delfolder = new File(newdir);
       
-      File[] delfolderList = delfolder.listFiles();
-      for(int i = 0; i< delfolderList.length ; i++) {
-         delfolderList[i].delete();
+      if(dto.getPdimg() != null) {
+    	  String newdir = uploadPath + "\\" + dto.getPdcode() + dto.getPdwriter();
+          File delfolder = new File(newdir);
+          File[] delfolderList = delfolder.listFiles();
+    	  for(int i = 0; i< delfolderList.length ; i++) {
+    	         delfolderList[i].delete();
+    	  }
+    	  delfolder.delete();
       }
       
-      delfolder.delete();
       return dao.delete(idx);
    }
